moved tests out of tests directory
